feat: idle-jpa 모듈로 분리
