+++ idle-jpa/src/main/java/com/ureca/idle/bookpreference/Hobulho.java
@@ -1,4 +1,4 @@
-package com.ureca.idle.idlejpa.bookpreference;
+package com.ureca.idle.bookpreference;
 
 import lombok.Getter;
