+++ idle-jpa/src/main/java/com/ureca/idle/submission/CurrentRoundSubmission.java
@@ -1,4 +1,4 @@
-package com.ureca.idle.idlejpa.submission;
+package com.ureca.idle.submission;
 
 import jakarta.persistence.*;
 import lombok.*;
@@ -8,13 +8,13 @@ import lombok.*;
 @AllArgsConstructor(access = AccessLevel.PRIVATE)
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @Entity
-public class PreviousRoundSubmission {
+public class CurrentRoundSubmission {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    @Column(nullable = false)
+    @Column(nullable = false, unique = true)
     private Long userId;
 
     @Column(nullable = false)
