+++ idle-jpa/src/main/java/com/ureca/idle/submission/PreviousRoundSubmission.java
@@ -1,0 +1,25 @@
+package com.ureca.idle.submission;
+
+import jakarta.persistence.*;
+import lombok.*;
+
+@Getter
+@Builder
+@AllArgsConstructor(access = AccessLevel.PRIVATE)
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@Entity
+public class PreviousRoundSubmission {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @Column(nullable = false)
+    private Long userId;
+
+    @Column(nullable = false)
+    private String name;
+
+    @Column(nullable = false)
+    private String phoneNum;
+}
