+++ idle-jpa/src/main/java/com/ureca/idle/user/User.java
@@ -1,5 +1,6 @@
-package com.ureca.idle.idlejpa.submission;
+package com.ureca.idle.user;
 
+import com.ureca.idle.config.BaseEntity;
 import jakarta.persistence.*;
 import lombok.*;
 
@@ -8,18 +9,25 @@ import lombok.*;
 @AllArgsConstructor(access = AccessLevel.PRIVATE)
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @Entity
-public class CurrentRoundSubmission {
+public class User extends BaseEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
     @Column(nullable = false, unique = true)
-    private Long userId;
+    private String email;
+
+    @Column(nullable = false)
+    private String password;
 
     @Column(nullable = false)
     private String name;
 
     @Column(nullable = false)
     private String phoneNum;
+
+    @Enumerated(EnumType.STRING)
+    @Column(nullable = false)
+    private Role role;
 }
